Moved unit tests for GetTaggedDecks() into Anki2DeckControllerTests
