+++ AnkiJapaneseFlashcardManagerTest/Anki2ControllerTests.cs
@@ -1,44 +1,11 @@
 using AnkiSentenceCardBuilder.Controllers;
 using AnkiSentenceCardBuilder.Models;
-using FluentAssertions;
 
 namespace AnkiJapaneseFlashcardManagerTest
 {
 	public class Anki2ControllerTests
 	{
 		string _anki2FolderPath = "./Resources/Anki2 Files/";
-
-		[Theory]
-		[InlineData("empty_kanjiResource_deck.anki2", "KanjiResource")]
-		[InlineData("empty_kanjiResource_newKanji_decks.anki2", "KanjiResource")]
-		[InlineData("empty_random_decks.anki2", "Random")]
-		[InlineData("empty_random_decks.anki2", "RandomResource")]
-		public void Get_decks_by_description_tag(string anki2File, string deckTagName)//TODO: Refactor to check for expected values
-		{
-			//Arange
-			Anki2Controller anki2Controller = new Anki2Controller(_anki2FolderPath + anki2File);
-
-			//Act
-			var taggedDecks = anki2Controller.GetTaggedDecks(deckTagName);
-
-			//Assert
-			taggedDecks.Should().NotBeEmpty();
-		}
-
-		[Theory]
-		[InlineData("empty_kanjiResource_deck.anki2", "NonExistentTag")]
-		[InlineData("empty_random_decks.anki2", "KanjiResource")]
-		public void No_decks_with_matching_description_tag_is_empty(string anki2File, string deckTagName)
-		{
-			//Arange
-			Anki2Controller anki2Controller = new Anki2Controller(_anki2FolderPath + anki2File);
-
-			//Act
-			var taggedDecks = anki2Controller.GetTaggedDecks(deckTagName);
-
-			//Assert
-			taggedDecks.Should().BeEmpty();
-		}
 
 		//[Theory]
 		//[InlineData("empty_kanjiResource_deck.anki2", 1706982246215)]
