+++ AnkiJapaneseFlashcardManagerTest/GlobalUsings.cs
@@ -1,1 +1,2 @@
 global using Xunit;
+global using FluentAssertions;
